# docs/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/community/stories": {
      "get": {
        "operationId": "CommunityController_list",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from a previous page.nextCursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "MEMORY",
                "LOCAL_STORY",
                "TRAVEL_EXPERIENCE",
                "THEN_AND_NOW",
                "DOCUMENT_CONTRIBUTION",
                "FAMILY_HISTORY",
                "PHOTO_STORY"
              ],
              "type": "string"
            }
          },
          {
            "name": "placeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "NEW",
                "HELPFUL"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "community"
        ]
      },
      "post": {
        "operationId": "CommunityController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommunityStoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "community"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/community/stories/mine": {
      "get": {
        "operationId": "CommunityController_listMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "community"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/community/stories/{slug}": {
      "get": {
        "operationId": "CommunityController_getBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "community"
        ]
      }
    },
    "/v1/community/stories/{slug}/comments": {
      "get": {
        "operationId": "CommunityController_getComments",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from a previous page.nextCursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "community"
        ]
      }
    },
    "/v1/community/stories/{id}": {
      "patch": {
        "operationId": "CommunityController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommunityStoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "community"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CommunityController_withdraw",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "community"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/community/stories/{id}/places": {
      "post": {
        "operationId": "CommunityController_linkPlace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkEntityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "community"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/community/stories/{id}/people": {
      "post": {
        "operationId": "CommunityController_linkPerson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkEntityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "community"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/community/stories/{id}/events": {
      "post": {
        "operationId": "CommunityController_linkEvent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkEntityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "community"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/community/stories/{id}/eras": {
      "post": {
        "operationId": "CommunityController_linkEra",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkEntityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "community"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/community/stories/{id}/vote": {
      "post": {
        "operationId": "CommunityController_vote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "community"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CommunityController_unvote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "community"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/community/stories/{id}/verification-state": {
      "patch": {
        "operationId": "CommunityController_setAuthorVerificationState",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAuthorVerificationStateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "community"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/community/stories/{id}/review-verification-state": {
      "patch": {
        "operationId": "CommunityController_setReviewVerificationState",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetReviewVerificationStateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "community"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/community/stories/{id}/moderation-status": {
      "patch": {
        "operationId": "CommunityController_setModerationStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetModerationStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "community"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Dau Viet API",
    "description": "Living Digital Atlas of Vietnam - backend API contract for web, mobile, and admin clients. See docs/backend/BACKEND_HANDOFF.md for the full handoff contract this document is generated alongside.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CommunityStoryTranslationInputDto": {
        "type": "object",
        "properties": {
          "locale": {
            "type": "string",
            "example": "vi"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        },
        "required": [
          "locale",
          "title"
        ]
      },
      "CreateCommunityStoryDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "MEMORY",
              "LOCAL_STORY",
              "TRAVEL_EXPERIENCE",
              "THEN_AND_NOW",
              "DOCUMENT_CONTRIBUTION",
              "FAMILY_HISTORY",
              "PHOTO_STORY"
            ]
          },
          "eventDateYear": {
            "type": "number",
            "description": "Only known component - never fabricate a month/day."
          },
          "eventDateMonth": {
            "type": "number"
          },
          "eventDateDay": {
            "type": "number"
          },
          "eventDatePrecision": {
            "type": "string",
            "enum": [
              "DAY",
              "MONTH",
              "YEAR",
              "DECADE",
              "CENTURY",
              "UNKNOWN"
            ]
          },
          "eventDateLabel": {
            "type": "string"
          },
          "heroMediaId": {
            "type": "string",
            "description": "Must be a MediaAsset the caller owns (spec section 16)."
          },
          "thenNowComparisonId": {
            "type": "string",
            "description": "Only valid when type = THEN_AND_NOW (spec section 18) - must be a ThenNowComparison the caller created."
          },
          "translations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommunityStoryTranslationInputDto"
            }
          }
        },
        "required": [
          "type",
          "translations"
        ]
      },
      "UpdateCommunityStoryDto": {
        "type": "object",
        "properties": {
          "eventDateYear": {
            "type": "number"
          },
          "eventDateMonth": {
            "type": "number"
          },
          "eventDateDay": {
            "type": "number"
          },
          "eventDatePrecision": {
            "type": "string",
            "enum": [
              "DAY",
              "MONTH",
              "YEAR",
              "DECADE",
              "CENTURY",
              "UNKNOWN"
            ]
          },
          "eventDateLabel": {
            "type": "string"
          },
          "heroMediaId": {
            "type": "string"
          },
          "translations": {
            "description": "Replaces the story's translations wholesale.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommunityStoryTranslationInputDto"
            }
          }
        }
      },
      "LinkEntityDto": {
        "type": "object",
        "properties": {}
      },
      "SetAuthorVerificationStateDto": {
        "type": "object",
        "properties": {}
      },
      "SetReviewVerificationStateDto": {
        "type": "object",
        "properties": {}
      },
      "SetModerationStatusDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}